pyproject.toml
[tool.poetry]
name = "seminario"
version = "0.3.0"
description = "Python package for seminar organization."
authors = ["Shota Imaki <shota.imaki@icloud.com>"]
license = "BSD-3-Clause"

[tool.poetry.dependencies]
python = "^3.6.1"
pandas = "^1.0.0"
pdfkit = "^0.6.1"
wkhtmltopdf = "^0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.3.2"
pytest-cov = "^2.8.1"
flake8 = "^3.7.9"
pyyaml = "^5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
